Rename hello package to prodeng via IDE
