--- src/main/java/ro/unibuc/hello/model/TodoEntity.java
@@ -1,12 +1,0 @@
-package ro.unibuc.hello.model;
-
-import org.springframework.data.annotation.Id;
-import org.springframework.data.mongodb.core.mapping.Document;
-
-@Document(collection = "todos")
-public record TodoEntity(
-    @Id String id,
-    String description,
-    boolean done,
-    String assignedUserId
-) {}
