--- src/main/java/ro/unibuc/hello/model/UserEntity.java
@@ -1,12 +1,0 @@
-package ro.unibuc.hello.model;
-
-import org.springframework.data.annotation.Id;
-import org.springframework.data.mongodb.core.mapping.Document;
-
-@Document(collection = "users")
-public record UserEntity(
-    @Id
-    String id,
-    String name,
-    String email
-) {}
