--- src/main/java/ro/unibuc/hello/request/AssignTodoRequest.java
@@ -1,11 +1,0 @@
-package ro.unibuc.hello.request;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-
-public record AssignTodoRequest(
-    @Email(message = "Invalid email format")
-    @NotBlank(message = "New assignee email is required")
-    String newAssigneeEmail
-) {}
-
