--- src/main/java/ro/unibuc/hello/request/ChangeNameRequest.java
@@ -1,8 +1,0 @@
-package ro.unibuc.hello.request;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record ChangeNameRequest(
-    @NotBlank(message = "Name is required")
-    String name
-) {}
