--- src/main/java/ro/unibuc/hello/request/CreateUserRequest.java
@@ -1,13 +1,0 @@
-package ro.unibuc.hello.request;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-
-public record CreateUserRequest(
-    @NotBlank(message = "Name is required")
-    String name,
-
-    @Email(message = "Invalid email format")
-    @NotBlank(message = "Email is required")
-    String email
-) {}
