--- src/main/java/ro/unibuc/hello/request/EditTodoRequest.java
@@ -1,8 +1,0 @@
-package ro.unibuc.hello.request;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record EditTodoRequest(
-    @NotBlank(message = "Description is required")
-    String description
-) {}
