--- src/main/java/ro/unibuc/hello/response/TodoResponse.java
@@ -1,9 +1,0 @@
-package ro.unibuc.hello.response;
-
-public record TodoResponse(
-    String id,
-    String description,
-    boolean done,
-    String assigneeName,
-    String assigneeEmail
-) {}
