--- src/main/java/ro/unibuc/hello/response/UserResponse.java
@@ -1,7 +1,0 @@
-package ro.unibuc.hello.response;
-
-public record UserResponse(
-    String id,
-    String name,
-    String email
-) {}
