--- src/main/java/ro/unibuc/hello/service/TodoService.java
@@ -1,99 +1,0 @@
-package ro.unibuc.hello.service;
-
-import java.util.List;
-import java.util.UUID;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import ro.unibuc.hello.model.TodoEntity;
-import ro.unibuc.hello.repository.TodoRepository;
-import ro.unibuc.hello.model.UserEntity;
-import ro.unibuc.hello.request.AssignTodoRequest;
-import ro.unibuc.hello.request.CreateTodoRequest;
-import ro.unibuc.hello.request.EditTodoRequest;
-import ro.unibuc.hello.response.TodoResponse;
-import ro.unibuc.hello.exception.EntityNotFoundException;
-
-@Service
-public class TodoService {
-
-    @Autowired
-    private TodoRepository todoRepository;
-
-    @Autowired
-    private UserService userService;
-
-    public List<TodoResponse> getTodosByUserEmail(String email) throws EntityNotFoundException {
-        UserEntity user = userService.getUserEntityByEmail(email);
-        List<TodoEntity> todos = todoRepository.findByAssignedUserId(user.id());
-        return todos.stream()
-                .map(todo -> toResponse(todo, user))
-                .toList();
-    }
-
-    public TodoResponse getTodoById(String id) throws EntityNotFoundException {
-        TodoEntity todo = todoRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException(id));
-        UserEntity assignee = userService.getUserEntityById(todo.assignedUserId());
-        return toResponse(todo, assignee);
-    }
-
-    public TodoResponse createTodo(CreateTodoRequest request) throws EntityNotFoundException {
-        UserEntity assignee = userService.getUserEntityByEmail(request.assigneeEmail());
-        TodoEntity todo = new TodoEntity(
-            UUID.randomUUID().toString(),
-            request.description(),
-            false,
-            assignee.id()
-        );
-        TodoEntity saved = todoRepository.save(todo);
-        return toResponse(saved, assignee);
-    }
-
-    public TodoResponse setDone(String id, boolean done) throws EntityNotFoundException {
-        TodoEntity existing = getEntityById(id);
-        TodoEntity updated = new TodoEntity(id, existing.description(), done, existing.assignedUserId());
-        TodoEntity saved = todoRepository.save(updated);
-        UserEntity assignee = userService.getUserEntityById(saved.assignedUserId());
-        return toResponse(saved, assignee);
-    }
-
-    public TodoResponse assign(String id, AssignTodoRequest request) throws EntityNotFoundException {
-        TodoEntity existing = getEntityById(id);
-        UserEntity newAssignee = userService.getUserEntityByEmail(request.newAssigneeEmail());
-        TodoEntity updated = new TodoEntity(id, existing.description(), existing.done(), newAssignee.id());
-        TodoEntity saved = todoRepository.save(updated);
-        return toResponse(saved, newAssignee);
-    }
-
-    public TodoResponse edit(String id, EditTodoRequest request) throws EntityNotFoundException {
-        TodoEntity existing = getEntityById(id);
-        TodoEntity updated = new TodoEntity(id, request.description(), existing.done(), existing.assignedUserId());
-        TodoEntity saved = todoRepository.save(updated);
-        UserEntity assignee = userService.getUserEntityById(saved.assignedUserId());
-        return toResponse(saved, assignee);
-    }
-
-    public void deleteTodo(String id) throws EntityNotFoundException {
-        if (!todoRepository.existsById(id)) {
-            throw new EntityNotFoundException(id);
-        }
-        todoRepository.deleteById(id);
-    }
-
-    private TodoEntity getEntityById(String id) throws EntityNotFoundException {
-        return todoRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException(id));
-    }
-
-    private TodoResponse toResponse(TodoEntity todo, UserEntity assignee) {
-        return new TodoResponse(
-            todo.id(),
-            todo.description(),
-            todo.done(),
-            assignee.name(),
-            assignee.email()
-        );
-    }
-}
