+++ src/main/java/ro/unibuc/prodeng/model/TodoEntity.java
@@ -1,0 +1,12 @@
+package ro.unibuc.prodeng.model;
+
+import org.springframework.data.annotation.Id;
+import org.springframework.data.mongodb.core.mapping.Document;
+
+@Document(collection = "todos")
+public record TodoEntity(
+    @Id String id,
+    String description,
+    boolean done,
+    String assignedUserId
+) {}
