+++ src/main/java/ro/unibuc/prodeng/model/UserEntity.java
@@ -1,0 +1,12 @@
+package ro.unibuc.prodeng.model;
+
+import org.springframework.data.annotation.Id;
+import org.springframework.data.mongodb.core.mapping.Document;
+
+@Document(collection = "users")
+public record UserEntity(
+    @Id
+    String id,
+    String name,
+    String email
+) {}
