+++ src/main/java/ro/unibuc/prodeng/repository/UserRepository.java
@@ -1,0 +1,13 @@
+package ro.unibuc.prodeng.repository;
+
+import java.util.Optional;
+
+import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.stereotype.Repository;
+import ro.unibuc.prodeng.model.UserEntity;
+
+@Repository
+public interface UserRepository extends MongoRepository<UserEntity, String> {
+
+    Optional<UserEntity> findByEmail(String email);
+}
