+++ src/main/java/ro/unibuc/prodeng/request/ChangeNameRequest.java
@@ -1,0 +1,8 @@
+package ro.unibuc.prodeng.request;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record ChangeNameRequest(
+    @NotBlank(message = "Name is required")
+    String name
+) {}
