+++ src/main/java/ro/unibuc/prodeng/request/CreateUserRequest.java
@@ -1,0 +1,13 @@
+package ro.unibuc.prodeng.request;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+
+public record CreateUserRequest(
+    @NotBlank(message = "Name is required")
+    String name,
+
+    @Email(message = "Invalid email format")
+    @NotBlank(message = "Email is required")
+    String email
+) {}
