+++ src/main/java/ro/unibuc/prodeng/request/EditTodoRequest.java
@@ -1,0 +1,8 @@
+package ro.unibuc.prodeng.request;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record EditTodoRequest(
+    @NotBlank(message = "Description is required")
+    String description
+) {}
