+++ src/main/java/ro/unibuc/prodeng/response/TodoResponse.java
@@ -1,0 +1,9 @@
+package ro.unibuc.prodeng.response;
+
+public record TodoResponse(
+    String id,
+    String description,
+    boolean done,
+    String assigneeName,
+    String assigneeEmail
+) {}
