+++ src/main/java/ro/unibuc/prodeng/response/UserResponse.java
@@ -1,0 +1,7 @@
+package ro.unibuc.prodeng.response;
+
+public record UserResponse(
+    String id,
+    String name,
+    String email
+) {}
