--- src/test/java/ro/unibuc/hello/e2e/CucumberSpringConfiguration.java
@@ -1,11 +1,0 @@
-package ro.unibuc.hello.e2e;
-
-import io.cucumber.spring.CucumberContextConfiguration;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@CucumberContextConfiguration
-@SpringBootTest(classes = CucumberSpringConfiguration.class)
-public class CucumberSpringConfiguration {
-    // Empty configuration - we don't need Spring context for E2E tests
-    // since we're calling an already running server via REST
-}
