+++ src/test/java/ro/unibuc/prodeng/e2e/HelloWorldE2E.java
@@ -1,0 +1,10 @@
+package ro.unibuc.prodeng.e2e;
+
+import io.cucumber.junit.Cucumber;
+import io.cucumber.junit.CucumberOptions;
+import org.junit.runner.RunWith;
+
+@RunWith(Cucumber.class)
+@CucumberOptions(features = "src/test/resources", tags = "E2E")
+public class HelloWorldE2E {
+}
