+++ src/test/java/ro/unibuc/prodeng/e2e/util/ResponseErrorHandler.java
@@ -1,0 +1,30 @@
+package ro.unibuc.prodeng.e2e.util;
+
+import org.springframework.http.client.ClientHttpResponse;
+
+import java.io.IOException;
+
+public class ResponseErrorHandler implements org.springframework.web.client.ResponseErrorHandler {
+
+    private ResponseResults results = null;
+    private Boolean hadError = false;
+
+    public ResponseResults getResults() {
+        return results;
+    }
+
+    public Boolean getHadError() {
+        return hadError;
+    }
+
+    @Override
+    public boolean hasError(ClientHttpResponse response) throws IOException {
+        hadError = response.getStatusCode().value() >= 400;
+        return hadError;
+    }
+
+    @Override
+    public void handleError(ClientHttpResponse response) throws IOException {
+        results = new ResponseResults(response);
+    }
+}
